Few changes
Ordered and commented  gui screens.
Stripped all the unused fonts from uGUI.
Replaced big font for temp display with UbuntuMono 45
Added file and line debug printing in case of Error_Handler call, and
set PWM emergency disable in that case.
This build should be pretty fail-safe and not burn the tip in case of
system error or lock-up.

Small fixes and improvements