Huge performance improvements.
Added widget content checking (new widget "last_value" variable).
Added widget and screen "force_update" variables.
All widgets will be drawed forcefully if screen "force_update" is set.
When a new screen is entered that bit is automatically set, and cleared
after all widgets have been drawn.

If the screen doesn't change, then the widget content checking is what
controls the drawing process.
The widget will not be drawed again if the content is the same.
In string widgets, the checking is done by suming up the string data,
and the result stored in last_value.

This releases a huge ammount of processing power, as before it was
updating all widgets in real time.

screen->force_refresh needs to be set when external code
disables/enables widgets, so the changes are drawed.

The main iron screen needed additional tweaking, as the iron temp font
is so big that partially overlaps existing widgets if drawed alone.
This was fixed by checking iron temp changes, and setting the force
refresh bit if different value is detected, so all the screen is drawn
when the iron temp value changes.

The result is almost 20x less cpu usage.
The main loop went from 108Hz to >1800Hz in the heaviest screen (Main
screen, drawing the big iron temp font is a heavy task).
Other screens reach up to 70KHz update.