Added active power limit based on supply voltage and heater impedance.
Specially done for lower impedance tips (C210, C245) that will take huge
powers at 24V.
This way low power tips can be used without risk of overpowering them.
Both settings can be adjustable in the iron settings.
The power applied to the tip will be constant independent on the power
supply voltage (as long as the voltage is enough).