Next attempt to avoid excessive EMA filter delay. Remove DEMA.
DEMA filtering seems not useful for this application, so it was removed.

The filter function is to reduce noise (small changes), but we don't
want it to filter huge changes, like when the tip is removed/plugged
back, or when the tip heats or cools very fast, as this adds delay and
can cause over/undershoot.

The typical noise is around 10-30 (ADC), so the approach was to compare
last average agains previous EMA value.
If Delta >200, totally reset the filter with the last data.
If Delta >100, add half the difference.

This greatly enhaces the response to fast changes, and the filtering
performance is unaffected.