Modified calibration screens.
Added checks to abort calibration if measured temp is >50ºC off the
target, so you can't burn the tip when doing the 450ºC calibration.

Added adjustment screen to manually set the default calibration values.
This doesn't need to be exact, just close enough so the calibration
process can work with them.
Everything is checked to ensure the entered data makes sense.
The save button will disabled if something is wrong.