Changed ADC and WAKE functions
Reading the NTC in the handle would cause a lot of noise when the heater
was tuned on and off, there was 0.2V drop in the gnd wire, so the temp
was always jumping few degrees up and down. Nobody noticed this before?
Now all the ADC channels are samples together in the PWM low time.

Changed WAKE reading method to polling.
The interrupt was getting triggered for every little noise, not really
needing that.