Move checksettings to settings.c
Delete settings saving flag.
Block irqs for only for each flash write, instead for all the length.
Enable Timer period elapsed interrupt for PWM timer.
Update the PWM in this moment.

Test: Read the PWM pin before and after the ADC conversion, if found to
be high, discard new readings and a make a beep to indicate the
detection.
Skip adc conversion if any error errors are found in the sequence.

Settings: Allow min save delay to 0 to allow disable saving. settings
this setting to 0 will never be saved to the flash, only for testing
purposes.
Change labels to "Stby time" and "Sleep time"