Oled dimmer changes.
Don't use the shake input for waking the screen, instead, detect station
mode.
If dimmer is enabled for all modes, only the encoder will wake up the
screen, or if the screen was off and the station exited sleep state.
If dimmer is enabled for low power modes modes, only the encoder or run
modes will wake up the screen.
So there won't be silly screen wakes.